refactor(terminals-home-agents): one home for the status vocabulary
T01. laneStatus.js now owns the words, the attention symbols and the small
formatters: one STATUS_LABELS table with a short-variant override, the new
ATTENTION_MARKS, and statusLabel/attentionMark/cleanCommand/
formatRelativeTime/assignmentIcon/assignmentText. laneBoard and
laneDetailRail read them from there; their local copies are gone, which
collapses STATUS_LABELS and STATUS_SHORT into one table and breaks the
rail's import dependency on the board.

statusLabel takes an options object so the board's long labels and the
rail's compressed ones are one function with two independent knobs. Output
is byte-identical at both call sites. _armQuietTimer and the init-once
guard untouched (C5).
