+++ .frame/specs/terminals-home-agents/status.json
@@ -1,7 +1,7 @@
 {
   "slug": "terminals-home-agents",
   "title": "Home, Terminals and Agents — three surfaces folded into one model",
-  "phase": "tasks_generated",
+  "phase": "implementing",
   "ai_tool": "claude-code",
   "generated_task_ids": [
     "task-spec-terminals-home-agents-T01",
@@ -16,7 +16,7 @@
     "task-spec-terminals-home-agents-T10"
   ],
   "created_at": "2026-08-26T10:51:12Z",
-  "updated_at": "2026-08-26T12:35:12.769Z",
-  "last_phase_at": "2026-08-26T12:25:32Z",
+  "updated_at": "2026-08-26T12:37:14.521Z",
+  "last_phase_at": "2026-08-26T12:37:14.522Z",
   "implement_mode": "autonomous"
 }
